+++ package.json
@@ -1,10 +1,10 @@
-{
+{               
   "name": "computed-columns",
   "version": "0.7.1",
   "description": "This visualization plugin is like a table, but with computed columns",
   "main": "index.js",
   "kibana": {
-    "version": "5.5.0"
+    "version": "kibana"
   },
   "scripts": {
     "lint": "eslint",
