+++ public/computed-columns-vis.js
@@ -1,151 +1,71 @@
-import _ from 'lodash';
-import { VisAggConfigProvider } from 'ui/vis/agg_config';
-import AggConfigResult from 'ui/vis/agg_config_result';
-import { AggResponseTabifyProvider } from 'ui/agg_response/tabify/tabify';
-import { uiModules } from 'ui/modules';
-import { Parser } from 'expr-eval';
-import numeral from 'numeral';
+import 'plugins/computed-columns/computed-columns.less';
+import 'plugins/computed-columns/computed-columns-vis-controller';
+import 'plugins/computed-columns/computed-columns-params';
+import mainTemplate from 'plugins/computed-columns/computed-columns-vis.html';
+import optionsTemplate from 'plugins/computed-columns/computed-columns-params.html';
 
-const module = uiModules.get('kibana/computed-columns', ['kibana']);
-module.controller('ComputedColumnsVisController', ($scope, $element, Private) => {
+import {CATEGORY} from 'ui/vis/vis_category';
+import {VisFactoryProvider} from 'ui/vis/vis_factory';
+import {VisTypesRegistryProvider} from 'ui/registry/vis_types';
+import {VisSchemasProvider} from 'ui/vis/editors/default/schemas';
 
-  const tabifyAggResponse = Private(AggResponseTabifyProvider);
-  const AggConfig = Private(VisAggConfigProvider);
+import image from './images/icon-table.svg';
 
-  const uiStateSort = ($scope.uiState) ? $scope.uiState.get('vis.params.sort') : {};
-  _.assign($scope.vis.params.sort, uiStateSort);
+VisTypesRegistryProvider.register(ExtendedMetricVisProvider);
 
-  const createExpressionsParams = (formula, row) => {
-    let regex = /col\[(\d+)\]/g;
-    let myArray;
-    let output = {};
-    while ((myArray = regex.exec(formula)) !== null) {
-      output[`x${myArray[1]}`] = (typeof row[myArray[1]].value === 'number') ?
-        numeral(row[myArray[1]].value).value() : row[myArray[1]].value;
-    }
-    return output;
-  };
+function ExtendedMetricVisProvider(Private) {
+  const VisFactory = Private(VisFactoryProvider);
+  const Schemas = Private(VisSchemasProvider);
 
-  const createParser = (computedColumn) => {
-    let expression = computedColumn.formula.replace(/col\[\d+\]/g, (value) => {
-      let cleanValue = /(\d+)/.exec(value)[1];
-      return `x${cleanValue}`;
-    });
-    return Parser.parse(expression);
-
-  };
-
-  const createColumn = (computedColumn, index) => {
-    let newColumn = {aggConfig: new AggConfig($scope.vis, {schema: 'metric', type: 'count'}), title: computedColumn.label};
-    newColumn.aggConfig.id = `1.computed-column-${index}`;
-    newColumn.aggConfig.key = `computed-column-${index}`;
-    return newColumn;
-  };
-
-  const createRows = (column, rows, computedColumn) => {
-    let parser = createParser(computedColumn);
-    return _.map(rows, (row) => {
-      let expressionParams = createExpressionsParams(computedColumn.formula, row);
-      let value = parser.evaluate(expressionParams);
-      let newCell = new AggConfigResult(column.aggConfig, void 0, value, value);
-      newCell.toString = () => {
-        return (typeof value === 'number') ? numeral(value).format(computedColumn.format) : value;
-      };
-      row.push(newCell);
-      return row;
-    });
-  };
-
-  const createTables = (tables, computedColumn, index) => {
-    _.forEach(tables, (table) => {
-      if (table.tables) {
-        createTables(table.tables, computedColumn, index);
-        return;
-      }
-
-      let newColumn = createColumn(computedColumn, index);
-      table.columns.push(newColumn);
-      table.rows = createRows(newColumn, table.rows, computedColumn);
-    });
-  };
-
-  const hideColumns = (tables, hiddenColumns) => {
-    if (!hiddenColumns) {
-      return;
-    }
-
-    let removedCounter = 0;
-    _.forEach(hiddenColumns.split(','), (item) => {
-      let index = item * 1;
-      _.forEach(tables, (table) => {
-        table.columns.splice(index - removedCounter, 1);
-        _.forEach(table.rows, (row) => {
-          row.splice(index - removedCounter, 1);
-        });
-      });
-      removedCounter++;
-    });
-  };
-
-  const shouldShowPagination = (tables, perPage) => {
-    return tables.some(function(table) {
-      if (table.tables) {
-        return shouldShowPagination(table.tables, perPage);
-      }
-      else {
-    	  return table.rows.length > perPage;
-      }
-    });
-  };
-
-  $scope.sort = $scope.vis.params.sort;
-  $scope.$watchCollection('sort', (newSort) => {
-    $scope.uiState.set('vis.params.sort', newSort);
-  });
-
-  $scope.$watchMulti(['esResponse', 'vis.params'], ([resp]) => {
-    console.debug('[computed-columns] Watch es response and vis params called');
-    let tableGroups = $scope.tableGroups = null;
-    let hasSomeRows = $scope.hasSomeRows = null;
-    let computedColumns = $scope.vis.params.computedColumns;
-    let hiddenColumns = $scope.vis.params.hiddenColumns;
-
-    if (resp) {
-      const vis = $scope.vis;
-      const params = vis.params;
-
-      tableGroups = tabifyAggResponse(vis, resp, {
-        partialRows: params.showPartialRows,
-        minimalColumns: vis.isHierarchical() && !params.showMeticsAtAllLevels,
-        asAggConfigResults: true
-      });
-
-      _.forEach(computedColumns, (computedColumn, index) => {
-        createTables(tableGroups.tables, computedColumn, index);
-      });
-
-      hideColumns(tableGroups.tables, hiddenColumns);
-
-      hasSomeRows = tableGroups.tables.some(function haveRows(table) {
-        if (table.tables) {
-          return table.tables.some(haveRows);
-        }
-        return table.rows.length > 0;
-      });
-
-      const showPagination = hasSomeRows && params.perPage && shouldShowPagination(tableGroups.tables, params.perPage);
-      $scope.tableVisContainerClass = {
-        'hide-pagination': !showPagination,
-        'hide-export-links': params.hideExportLinks
-      };
-
-      $element.trigger('renderComplete');
-    }
+  return VisFactory.createAngularVisualization({
+    name: 'computed-columns',
+    title: 'Computed Cols',
+    icon: 'fa-table',
+    description: 'Same functionality than Data Table, but after data processing, computed columns can be added with math expressions.',
+    category: CATEGORY.DATA,
     
-    $scope.hasSomeRows = hasSomeRows;
-    if (hasSomeRows) {
-      $scope.tableGroups = tableGroups;
+    //visualization: VisController,
+
+    visConfig: {
+      defaults: {
+        perPage: 10,
+        showPartialRows: false,
+        showMeticsAtAllLevels: false,
+        sort: {
+          columnIndex: null,
+          direction: null
+        },
+        showTotal: false,
+        totalFunc: 'sum',
+        computedColumns: [{
+          formula: 'col[0] * col[0]',
+          label: 'Value squared',
+          format: '0,0.[00]',
+          enabled: true
+        }],
+        hideExportLinks: false
+      },
+      template: mainTemplate,
+    },
+    editorConfig: {
+      optionsTemplate: optionsTemplate,
+      schemas: new Schemas([{
+        group: 'metrics',
+        name: 'metric',
+        title: 'Metric',
+        min: 1,
+      }, {
+        group: 'buckets',
+        name: 'bucket',
+        title: 'Split Rows'
+      }, {
+        group: 'buckets',
+        name: 'split',
+        title: 'Split Table'
+      }]),
     }
   });
+}
 
-});
+export default ExtendedMetricVisProvider;
+
